+++ src/gui/plugins/component_inspector/CMakeLists.txt
@@ -1,10 +1,4 @@
 gz_add_gui_plugin(ComponentInspector
-  SOURCES
-    AirPressure.cc
-    ComponentInspector.cc
-    Types.cc
-  QT_HEADERS
-    AirPressure.hh
-    ComponentInspector.hh
-    Types.hh
+  SOURCES ComponentInspector.cc
+  QT_HEADERS ComponentInspector.hh
 )
