+++ src/gui/plugins/component_inspector/ComponentInspector.qml
@@ -78,32 +78,13 @@ Rectangle {
     return _model.dataType + '.qml'
   }
 
-  // Get number of decimal digits based on a width value
-  // \param[in] _width Pixel width
-  // \return Number of decimals that fit with the provided width.
+  // Get number of decimal digits based on a widget's width
   function getDecimals(_width) {
-    // Use full decimals if the width is <= 0, which allows the value
-    // to appear correctly.
-    if (_width <= 0 || _width > 110)
-      return 6;
-
     if (_width <= 80)
       return 2;
-
-    return 4;
-  }
-
-  // Get number of decimal digits based on a widget's width, and adjust the
-  // widget's value to prevent zero padding.
-  // \param[in, out] _widgetId The widget id that will display the value. This
-  // widget's width attribute is used to determine the number of decimals.
-  // \param[in] _value The value that should be used to set the _widgetId's
-  // value attribute.
-  function getDecimalsAdjustValue(_widgetId, _value) {
-    // Make sure to update the value, otherwise zeros are used intead of 
-    // the actual values.
-    _widgetId.value = _widgetId.activeFocus ? _widgetId.value : _value;
-    return getDecimals(_widgetId.width);
+    else if (_width <= 100)
+      return 4;
+    return 6;
   }
 
   /**
