--- src/gui/plugins/component_inspector/Types.cc
@@ -1,33 +1,0 @@
-/*
- * Copyright (C) 2021 Open Source Robotics Foundation
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
-*/
-
-#include "Types.hh"
-
-void ignition::gazebo::setNoise(sdf::Noise &_noise,
-    double _mean, double _meanBias, double _stdDev,
-    double _stdDevBias, double _dynamicBiasStdDev,
-    double _dynamicBiasCorrelationTime)
-{
-  _noise.SetMean(_mean);
-  _noise.SetBiasMean(_meanBias);
-
-  _noise.SetStdDev(_stdDev);
-  _noise.SetBiasStdDev(_stdDevBias);
-
-  _noise.SetDynamicBiasStdDev(_dynamicBiasStdDev);
-  _noise.SetDynamicBiasCorrelationTime(_dynamicBiasCorrelationTime);
-}
