+++ src/gui/plugins/modules/TypeIcon.qml
@@ -44,7 +44,6 @@ Image {
         entityType == 'joint' ||
         entityType == 'light' ||
         entityType == 'link' ||
-        entityType == 'sensor' ||
         entityType == 'model')
     {
       entityImage = entityType
