+++ src/gui/resources/gazebo.qrc
@@ -10,7 +10,6 @@
     <file>images/minus.png</file>
     <file>images/model.png</file>
     <file>images/plus.png</file>
-    <file>images/sensor.png</file>
     <file>images/unlock.svg</file>
     <file>images/visual.png</file>
   </qresource>
